Added transaction details and history navigation
In the transactions page, transaction details can now be viewed in a dedicated modal. A button to navigate to the transaction's history page has also been added within this detailed view. In addition, the route parameter for the transaction history page has been updated to use the transaction name instead of id.